requirements.txt
# Core dependencies
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Enhanced UI components
plotly>=5.17.0
streamlit-lottie>=0.0.5
streamlit-extras>=0.3.5
streamlit-option-menu>=0.3.6

# Audio processing
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
pyaudio>=0.2.11
audio-recorder-streamlit>=0.0.8

# Speech recognition
openai-whisper>=20231117
speechrecognition>=3.10.0

# NLP and medical entity extraction
transformers>=4.35.0
torch>=2.1.0
spacy>=3.7.0
scispacy>=0.5.3
medspacy>=1.0.0

# Data processing
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0

# PDF generation
reportlab>=4.0.0
fpdf2>=2.7.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.10.0
flake8>=6.1.0

# Optional: For production deployment
gunicorn>=21.2.0
